feat!: Replace `PsbtUpdater` and `Builder` with `create_psbt` method
**Remove `PsbtUpdater` & `Builder`**

These was no need for these to be separated. The builder pattern was
nice but not needed at this stage.

**Remove `DataProvider`**

Instead of creating a separate struct for passing in data for
inputs/outputs. We can just have the data within our inputs/outputs.

**Introduce `Input` and `Output` types**

The `Input` type is what is passed through all stages of tx-building.
From canonicalization, grouping/filtering, coin-selection and finally
used to create the psbt.

The `Output` type is used to describe a psbt output. It either is one
that we own (in which, we have the descriptor), or one we do not (in
which case, we just have a spk).