feat!: Remove `bdk_chain` dependency
`CanonicalUnspents` is our "single source of truth" and all UTXOs added
to the selector should really pass through it to ensure we don't create
transactions that double-spend themselves.

Also replaced `Input::from_finalized_psbt_input` with `from_psbt_input`
for more flexibility.

The responsibility of grouping and filtering input candidates have been
moved to `InputCandidates` (instead of the now-removed `CoinControl`
struct).

The downside of this is that it makes it more invloved to interface with
`bdk_chain` or `bdk_wallet`. Adding methods on `KeychainTxOutIndex` will
help.