feat(selector)!: allow multiple change sources
Previously, the `change_descriptor` field only accepted a definite
descriptor to derive the change output script pubkey. This worked for
most cases but failed for scenarios - such as silent payments - where a
definite descriptor is not available.

Now `change_descriptor` is `change_script` and can take a ScriptSource,
which can be a descriptor or the script itself.

Also, `change_weights` field is added as a field of SelectorParams, to
externalize the calculus of the satisfaction weight from the API.

This allowed the remove of the `to_cs_change_weigths` method.

`From<(ScriptSource, Amount)>` for Output is added for conveniency.